requirements.txt
black
bs4
python-dateutil
requests
tqdm